Create Index, Style and Script
Created the main idea for the pages in the index and style.
Matches the idea on the sign up form for Odin but a few
differences. For one, I added a progress bar at the bottom
of the main form. I styled the input fields differently
too. Next I want to make the progress bar functional
and I want to see if I can create a map for country codes
in the number section. The user would have a drop down menu
of coutnry code and the pattern to match would be different
for each code based on E.164 standard.